DAG-760: Små justering på spacing, oppdatering etter kommentarer på pul request
