+++ src/components/account-number/AccountNumber.module.css
@@ -1,13 +1,8 @@
 .accountNumberContainer {
-  display: grid;
   padding: var(--a-spacing-4);
   border-radius: var(--a-border-radius-large);
   background-color: var(--a-gray-50);
   font-weight: var(--a-font-weight-bold);
-}
-
-.textContainer {
-  margin: var(--a-spacing-4) 0 var(--a-spacing-3);
 }
 
 .accountNumber {
