+++ src/components/meld-fra-om-endring/MeldFraOmEndring.module.css
@@ -15,6 +15,12 @@
   margin-top: var(--a-spacing-6);
 }
 
+/* Override background color for secondary button, known on design system */
+/* This should be removed after design system has fixed this bug */
+.navigationContainer a {
+  background-color: transparent;
+}
+
 @media screen and (min-width: 700px) {
   .navigationContainer {
     grid-template-columns: 1fr 1fr;
