+++ src/components/section/Section.module.css
@@ -3,6 +3,10 @@
   width: 100%;
 }
 
+.smallSpacing.smallSpacing {
+  padding: var(--a-spacing-4);
+}
+
 .gridContainer {
   display: grid;
   grid-template-columns: 1fr;
@@ -17,7 +21,7 @@
 @media screen and (min-width: 700px) {
   .section {
     margin: 0 auto;
-    padding: var(--a-spacing-10) var(--a-spacing-4);
+    padding: var(--a-spacing-10) var(--a-spacing-4) var(--a-spacing-8);
     max-width: 100vw;
   }
 
@@ -26,12 +30,12 @@
   }
 
   /* 66.66% width */
-  .defaultRowWidth {
+  .defaultWidth {
     grid-column-start: span 2;
   }
 
   /* 100% width */
-  .fullRowWidth {
+  .fullWidth {
     grid-column-start: span 3;
   }
 }
