+++ src/components/section/Section.tsx
@@ -5,15 +5,17 @@ import styles from "./Section.module.css";
 interface IProps {
   highlighted?: boolean;
   children: ReactNode;
+  smallSpacing?: boolean;
 }
 
 export function Section(props: IProps) {
-  const { highlighted, children } = props;
+  const { highlighted, children, smallSpacing } = props;
 
   return (
     <section
       className={classnames(styles.section, {
         [styles.highlighted]: highlighted,
+        [styles.smallSpacing]: smallSpacing,
       })}
     >
       {children}
